requirements.txt
# Dipendenze principali
PySide6>=6.0.0
opencv-python>=4.5.0
numpy>=1.19.0

# Dipendenze di rete
pyzmq>=22.0.0
zeroconf>=0.38.0

# Utilità
Pillow>=8.0.0
scipy>=1.7.0
psutil>=5.8.0

# Per la decodifica video
simplejpeg>=1.6.0

# Strumenti di sviluppo (opzionali)
pytest>=6.0.0
black>=21.5b2
flake8>=3.9.0